Move type(...) code into type_call.rs
